refactor: reestructurando tareas de scrapping y creando nueva tarea para notificaciones
se separan las tareas de base de datos y scrapping y se crea una tarea para enviar notificaciones de
forma manual
