+++ .github/workflows/scrapping.yml
@@ -1,4 +1,4 @@
-name: Playwright Tests
+name: Scrapping
 on:
   schedule:
     - cron: "30/5 20 * * *"
@@ -21,7 +21,7 @@ jobs:
     - name: Install Playwright Browsers
       run: npx playwright install --with-deps
     - name: Run Playwright tests
-      run: npx playwright test
+      run: npx playwright test scraping.spec.ts
     - uses: actions/upload-artifact@v4
       if: ${{ !cancelled() }}
       with:
