+++ __e2e__/notification.spec.ts
@@ -1,0 +1,7 @@
+import { test } from "@playwright/test";
+
+import { sendNotification } from "@/tasks/notification";
+
+test("should send the notifications", async () => {
+	await sendNotification();
+});
