+++ package.json
@@ -8,7 +8,8 @@
 		"start": "next start",
 		"lint": "next lint",
 		"test:unit": "vitest --dir=./tests",
-		"test:e2e": "npx playwright test",
+		"test:e2e:scrapping": "npx playwright test scraping.spec.ts",
+		"test:e2e:notification": "npx playwright test notification.spec.ts",
 		"drizzle:generate": "drizzle-kit generate",
 		"drizzle:migrate": "drizzle-kit migrate",
 		"drizzle:push": "drizzle-kit push",
